Add files via upload
finition tp3

+++ src/main/java/fr/diginamic/Spring_Data_JPA/repository/AnimalRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import fr.diginamic.Spring_Data_JPA.model.Animal;
 
 @Repository
-public interface AnimalRepository extends CrudRepository<Animal, Long> {
+public interface AnimalRepository extends CrudRepository<Animal, Integer> {
     
 }
