+++ src/main/java/fr/diginamic/Spring_Data_JPA/repository/SpeciesRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import fr.diginamic.Spring_Data_JPA.model.Species;
 
 @Repository
-public interface SpeciesRepository extends CrudRepository<Species, Long> {
+public interface SpeciesRepository extends CrudRepository<Species, Integer> {
     
 }
